backend/requirements.txt
absl-py==2.2.2
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
-e git+https://github.com/guilleps/vizel-art-similarity.git@d01c18616c8879400339bf029546cc59fbabc986#egg=backend&subdirectory=backend\backend
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.44.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
dotenv==0.9.9
drf-spectacular==0.28.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
itypes==1.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
namex==0.0.8
numpy==2.1.3
openapi-codec==1.3.2
optree==0.15.0
packaging==24.2
pillow==11.2.1
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
protobuf==5.29.4
protoc-gen-openapiv2==0.0.1
psycopg2==2.9.10
Pygments==2.19.1
pytest==8.4.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.3
setuptools==79.0.0
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
termcolor==3.0.1
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
